feat: create custom hook to add search params
