+++ src/components/SearchInput.jsx
@@ -1,20 +1,9 @@
 import React, { useState } from "react";
-import { useHistory, useLocation } from "react-router-dom";
+import { useSearchParams } from '../hooks/useSearchParams';
 
 const SearchInput = ({ page, query }) => {
-  const { push } = useHistory();
-  const { search } = useLocation();
   const [searchValue, setSearchValue] = useState(query ? query : "");
-
-  const addQuery = (key, value) => {
-    const searchParams = new URLSearchParams(search); 
-    searchParams.set(key, value);
-    searchParams.delete('page');
-    push({
-      pathname: 'artworks',
-      search: searchParams.toString()
-    });
-  };
+  const { updateQueryParams } = useSearchParams();
 
   const searchMargin = page === "home" ? "lg:mx-7" : "lg:ml-7";
 
@@ -22,7 +11,7 @@ const SearchInput = ({ page, query }) => {
 
   const submitHandler = (e) => {
     e.preventDefault();
-    addQuery('q', searchValue.toLowerCase());
+    updateQueryParams('q', searchValue.toLowerCase());
   };
 
   return (
